captains-form-api-modified

+++ v1-adminaccess/routes/OurShipCaptions.mjs
@@ -2,35 +2,54 @@
 import express from 'express';
 import { client } from '../../db/mongodb.mjs'
 import { ObjectId } from 'mongodb'
+import multer from 'multer'
+import path from "path"
+const __dirname = path.resolve()
+import fs from "fs"
 
 const db = client.db("yacht");
 const col = db.collection("shipcaption");
 
 let router = express.Router()
+const storage = multer.diskStorage({
+    destination: (req, file, cb) => {
+        cb(null, './uploads/captainsImages'); // Uploads will be stored in the 'uploads' directory
+    },
+    filename: (req, file, cb) => {
+        cb(null, `${Date.now()}-${file.originalname}`);
+    },
+});
+
+const upload = multer({ storage });
 
 
 
 // POST    /api/v1/post
-router.post('/add-ship-caption', async (req, res, next) => {
+router.post('/add-ship-caption', upload.single('head-image'),  async (req, res, next) => {
 
 
     console.log(req.body)
-    const {image , name , email , Instagramlink , twitterink , facebooklink , linkedinlink , phonenumber} = req.body
+    const { image, name, email, instagramLink, twitterLink, facebookLink, linkedinlink, phonenumber } = req.body
     try {
-
+        if (!req.file) {
+            return res.status(400).send("No file uploaded")
+        }
+        const uploadedFile = req.file;
+        const imagePath = uploadedFile.path;
         const socialMediaLinks = {
-            Instagram:Instagramlink,
-            fascebook:facebooklink,
-            linkedin:linkedinlink,
-            twitter:twitterink,
+            Instagram: instagramLink,
+            fascebook: facebookLink,
+            linkedin: linkedinlink,
+            twitter: twitterLink,
         }
         const insertResponse = await col.insertOne({
             // _id: "7864972364724b4h2b4jhgh42",
-            image:image,
+            image: image,
             Name: name,
             phonenumber: phonenumber,
             email: email,
-            socialMediaLinks
+            socialMediaLinks,
+            image:imagePath
         });
         console.log("insertResponse: ", insertResponse);
 
@@ -104,8 +123,16 @@ router.delete('/ship-caption/:postId', async (req, res, next) => {
     }
 
     try {
+        const post = await col.findOne({_id : new ObjectId(req.params.postId)})
+        const filePath = post.image; 
+        console.log(filePath)
+    if (filePath) {
+        // Delete the file from the server folder
+        const imagePath = path.join(__dirname,  filePath);
+        fs.unlinkSync(imagePath);
+      }
         const deleteResponse = await col.findOneAndDelete({ _id: new ObjectId(req.params.postId) });
-        
+
         deleteResponse ? res.send('post deleted') : res.status(404).send('no post found')
         console.log("deleteResponse: ", deleteResponse);
     } catch (e) {
